refactor: move bus inmmemory to kit
