+++ kit/bus/inmemory/command.go
