+++ kit/bus/inmemory/event.go
